chore: Move 'General Menu' and 'Network Menu' to own files
