+++ source/components/Header/Menus/index.ts
@@ -1,0 +1,2 @@
+export { GeneralMenu } from './GeneralMenu';
+export { NetworkMenu } from './NetworkMenu';
